Moved pom_ files into pom package
